% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgs_explore.R
\name{bgs_explore}
\alias{bgs_explore}
\title{bgs_explore}
\usage{
bgs_explore(x)
}
\arguments{
\item{x}{List object as read from read_bgs(...)}
}
\value{
Tibble with variable names and their total occurrences.
}
\description{
Quick way to explore underlying BGStats data in all its chaotic energy.
}
