% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgs_games.R
\name{bgs_games}
\alias{bgs_games}
\title{bgs_games}
\usage{
bgs_games(x)
}
\arguments{
\item{x}{List object as read from read_bgs(...)}
}
\value{
Tibble with player information.
}
\description{
Extracts game information and flags expansions.
}
