% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgs_players.R
\name{bgs_players}
\alias{bgs_players}
\title{bgs_players}
\usage{
bgs_players(x)
}
\arguments{
\item{x}{List object as read from read_bgs(...)}
}
\value{
Tibble with player information.
}
\description{
Extracts player information and flags AI players.
}
