% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgs_plays.R
\name{bgs_plays}
\alias{bgs_plays}
\title{bgs_plays}
\usage{
bgs_plays(x, idx_solo = NULL)
}
\arguments{
\item{x}{List object as read from read_bgs(...)}

\item{idx_solo}{Possible vector to}
}
\value{
Tibble with full games history.
}
\description{
Extracts full play history
}
