% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbgs.R
\name{ggbgs}
\alias{ggbgs}
\title{ggbgs}
\usage{
ggbgs(
  x,
  y = c("hrs", "n"),
  game = NULL,
  yr_min = NULL,
  yr_max = NULL,
  origin = 1,
  dt = c("month", "year"),
  prop = FALSE,
  pal = c("#313131", "#F28F1C")
)
}
\arguments{
\item{x}{List object as read from read_bgs(...)}

\item{y}{Whether to plot play count (n) or number of hours played (hrs)}

\item{yr_min}{Minimum year}

\item{yr_max}{Maximum year}

\item{origin}{Starting month for x axis.}

\item{dt}{Note to summarise by month or year}

\item{prop}{Logical on whether to plot y axis as raw values or percentages.}

\item{pal}{Colours to use for plotting}
}
\value{
Tibble with player information.
}
\description{
Plots boardgame history.
}
