% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bgs.R
\name{read_bgs}
\alias{read_bgs}
\title{read_bgs}
\usage{
read_bgs(x)
}
\arguments{
\item{x}{.json file path.}
}
\value{
list object storing all the bells and whistles from a .json BGStats object.
}
\description{
Reads JSON files as exported from the BGStats app.
}
